feat: finish jest client test
