+++ package.json
@@ -31,12 +31,12 @@
     "jest": "^29.1.2",
     "prettier": "^2.6.2",
     "ts-jest": "^29.0.3",
-    "ts-node-dev": "^1.1.8",
+    "ts-node-dev": "^2.0.0",
     "typescript": "^4.6.4"
   },
   "scripts": {
     "build": "tsc",
-    "dev": "ts-node-dev -r tsconfig-paths/register --transpile-only --ignore-watch node_modules src/shared/http/server.ts",
+    "dev": "ts-node-dev -r tsconfig-paths/register --transpile-only --ignore-watch node_modules src/shared/infra/http/server.ts",
     "test": "jest"
   }
 }
