+++ src/shared/container/index.ts
@@ -1,5 +1,5 @@
 import {container} from 'tsyringe';
-import ClientRepository from '@modules/clients/repositories/client-repository';
+import ClientRepository from '@modules/clients/infra/repositories/client-repository';
 import IClientRepository from '@modules/clients/repositories/IClient-repositoy';
