+++ src/shared/infra/http/app.ts
@@ -3,8 +3,8 @@ import clientRouter from '@modules/clients/infra/http/routes';
 import '@shared/container';
 import express from 'express';
 import cors from 'cors';
-import { ErrorHandler } from './../middleware/errorHandler';
-import dbConnection from '../../mongoose/dbConnection';
+import { ErrorHandler } from './middleware/errorHandler';
+import dbConnection from './dbConnection';
 import 'express-async-errors';
 
 dbConnection();
