+++ src/shared/infra/http/server.ts
@@ -1,5 +1,5 @@
 import app from './app';
-import '../../../'
+
 app.listen(process.env.PORT || 3333, () => {
     console.log('⚡️ Server listening on port 3333');
   });
